pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lark-to-notes"
version = "0.1.0"
description = "Ingest Lark messages into an Obsidian notes vault"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
dependencies = [
    "pyyaml>=6.0.3",
    "structlog>=25.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.15",
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "ruff>=0.11",
]

[project.scripts]
lark-to-notes = "lark_to_notes.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/lark_to_notes"]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "ANN", # flake8-annotations
    "S",   # flake8-bandit
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific rules
]
ignore = [
    "ANN401", # Any allowed for dynamic dispatch
    "S101",   # use of assert (ok in tests)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S", "E501", "TCH"]
"automation/**" = ["ANN", "N", "UP", "B", "SIM", "RUF", "E501"]
"src/lark_to_notes/storage/db.py" = ["TCH003"]
"src/lark_to_notes/intake/ledger.py" = ["TCH003"]
"src/lark_to_notes/intake/replay.py" = ["TCH003"]
"src/lark_to_notes/render/models.py" = ["TCH003"]
"src/lark_to_notes/render/writer.py" = ["TCH003"]
"src/lark_to_notes/runtime/lock.py" = ["TCH003"]
"src/lark_to_notes/runtime/reconcile.py" = ["TCH003"]
"src/lark_to_notes/runtime/registry.py" = ["TCH003"]

[tool.mypy]
python_version = "3.12"
strict = true
files = ["src", "tests"]
exclude = ["automation/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
pythonpath = ["."]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20260408",
]
